2d-render/pyproject.toml
[project]
name = "2davatars2drender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "colored>=2.3.1",
    "cuda-python==12.9.0",
    "cython>=3.1.2",
    "einops>=0.8.1",
    "filetype>=1.2.0",
    "grpcio==1.74.0",
    "grpcio-tools==1.74.0",
    "imageio>=2.37.0",
    "imageio-ffmpeg>=0.6.0",
    "kornia>=0.8.1",
    "librosa>=0.11.0",
    "loguru>=0.7.3",
    "numpy==2.0.1",
    "opencv-python-headless>=4.12.0.88",
    "pillow>=11.3.0",
    "polygraphy>=0.49.26",
    "scikit-image>=0.25.2",
    "timm>=1.0.19",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "transformers>=4.54.0",
    "uuid>=1.30",
]
